+++ ChangeLog.md
@@ -23,6 +23,7 @@
     (now using the Myers diff algorithm)
 - Added syntax highlighting for flags
 - Sound change rules can now be specified on the command-line using new `--eval` or `-e` flag
+- Add CLI option to highlight words different to input
 
 ### Code